Global snapshots new flow
