+++ kotsadm/web/src/Root.jsx
@@ -7,7 +7,8 @@ import Modal from "react-modal";
 import find from "lodash/find";
 import ConnectionTerminated from "./ConnectionTerminated";
 import GitOps from "././components/clusters/GitOps";
-import SnapshotSettings from "./components/snapshots/Snapshots";
+import Snapshots from "./components/snapshots/Snapshots";
+import SnapshotSettings from "./components/snapshots/SnapshotSettings";
 import SnapshotDetails from "./components/snapshots/SnapshotDetails";
 import PreflightResultPage from "./components/PreflightResultPage";
 // import Redactors from "./components/redactors/Redactors";
@@ -153,7 +154,7 @@ class Root extends Component {
         rootDidInitialWatchFetch: true
       });
       return apps;
-    } catch(err) {
+    } catch (err) {
       throw err;
     }
   }
@@ -259,12 +260,13 @@ class Root extends Component {
     const apps = this.state.appsList;
     return !!find(apps, app => app.allowSnapshots);
   }
-  
+
   onLogoutError = (message) => {
     this.setState({
       errLoggingOut: message
     })
   }
+
 
   render() {
     const {
@@ -324,8 +326,9 @@ class Root extends Component {
                   <Route path="/unsupported" component={UnsupportedBrowser} />
                   <ProtectedRoute path="/cluster/manage" render={(props) => <ClusterNodes {...props} appName={this.state.selectedAppName} />} />
                   <ProtectedRoute path="/gitops" render={(props) => <GitOps {...props} appName={this.state.selectedAppName} />} />
-                  <ProtectedRoute path="/snapshots" render={(props) => <SnapshotSettings {...props} appName={this.state.selectedAppName} />} />
-                  <ProtectedRoute path="/snapshot/details" render={(props) => <SnapshotDetails {...props} appName={this.state.selectedAppName} />} />
+                  <ProtectedRoute exact path="/snapshots" render={(props) => <Snapshots {...props} appName={this.state.selectedAppName} />} />
+                  <ProtectedRoute exact path="/snapshots/settings" render={(props) => <SnapshotSettings {...props} appName={this.state.selectedAppName} />} />
+                  <ProtectedRoute exact path="/snapshots/details/:id" render={(props) => <SnapshotDetails {...props} appName={this.state.selectedAppName} />} />
                   {/* <ProtectedRoute exact path="/redactors" render={(props) => <Redactors {...props} appName={this.state.selectedAppName} />} />
                   <ProtectedRoute exact path="/redactors/new" render={(props) => <EditRedactor {...props} appName={this.state.selectedAppName} isNew={true} />} />
                   <ProtectedRoute exact path="/redactors/:slug" render={(props) => <EditRedactor {...props} appName={this.state.selectedAppName} />} /> */}
