+++ kotsadm/web/src/components/snapshots/SnapshotRow.jsx
@@ -1,5 +1,6 @@
 
 import React from "react";
+import { Link } from "react-router-dom";
 import ReactTooltip from "react-tooltip"
 import dayjs from "dayjs";
 import isSameOrAfter from "dayjs/plugin/isSameOrAfter";
@@ -8,7 +9,7 @@ dayjs.extend(isSameOrAfter);
 import { Utilities } from "../../utilities/utilities";
 
 
-class SnapshotDetailsRow extends React.Component {
+class SnapshotRow extends React.Component {
 
   handleDeleteClick = snapshot => {
     this.props.toggleConfirmDeleteModal(snapshot);
@@ -24,16 +25,17 @@ class SnapshotDetailsRow extends React.Component {
 
     return (
       <div className={`flex flex-auto SnapshotRow--wrapper alignItems--center ${snapshot?.status === "Deleting" && "is-deleting"} ${isExpired && "is-expired"}`}>
-        <div className="flex-column flex1">
-          <div className="flex flex-column">
-            <p className={`u-fontSize--largest ${isExpired || snapshot?.status === "Deleting" ? "u-color--dustyGray" : "u-color--tuna"} u-lineHeight--normal u-fontWeight--bold u-marginRight--10`}>{snapshot?.name}</p>
-            <div className="flex flex1 u-marginTop--10">
-              <p className="u-fontSize--small u-color--dustyGray u-fontWeight--medium u-lineHeight--normal u-marginRight--20">{snapshot?.startedAt ? Utilities.dateFormat(snapshot?.startedAt, "MMM D YYYY @ hh:mm a") : "n/a"}</p>
+        <div className="flex-column flex1" style={{maxWidth: "700px"}}>
+          <p className={`u-fontSize--largest ${isExpired || snapshot?.status === "Deleting" ? "u-color--dustyGray" : "u-color--tuna"} u-lineHeight--normal u-fontWeight--bold u-marginRight--10`}>{snapshot?.name}</p>
+          <div className="flex flex1 alignItems--center u-marginTop--10">
+            <p className="u-fontSize--small u-color--dustyGray u-fontWeight--medium u-lineHeight--normal u-marginRight--20">{snapshot?.startedAt ? Utilities.dateFormat(snapshot?.startedAt, "MMM D YYYY @ hh:mm a") : "n/a"}</p>
+            {snapshot?.status === "Completed" ?
               <p className="u-fontSize--small u-color--dustyGray u-fontWeight--medium u-lineHeight--normal u-marginRight--20">
                 <span className={`status-indicator u-marginRight--5 ${snapshot?.status.toLowerCase()}`}>{Utilities.snapshotStatusToDisplayName(snapshot?.status)}</span>
                 on {snapshot?.finishedAt ? (snapshot?.finishedAt ? Utilities.dateFormat(snapshot?.finishedAt, "MMM D YYYY @ hh:mm a") : "TBD") : "n/a"}
-              </p>
-            </div>
+              </p> :
+              <span className={`status-indicator u-marginRight--5 ${snapshot?.status.toLowerCase()}`}>{Utilities.snapshotStatusToDisplayName(snapshot?.status)}</span>
+            }
           </div>
         </div>
         <div className="flex flex1">
@@ -48,12 +50,12 @@ class SnapshotDetailsRow extends React.Component {
           <div className="flex flex-auto">
             {snapshot?.status === "Completed" &&
               <div className="flex">
-                <span className="icon snapshot-restore-icon u-cursor--pointer" onClick={() => this.handleRestoreClick(snapshot)} data-tip="Restore from this backup"/>
+                <span className="icon snapshot-restore-icon u-cursor--pointer" onClick={() => this.handleRestoreClick(snapshot)} data-tip="Restore from this backup" />
                 <ReactTooltip effect="solid" className="replicated-tooltip" />
               </div>}
             {snapshot?.status !== "InProgress" &&
               <span className="icon snapshot-trash-icon u-marginLeft--20 u-cursor--pointer" onClick={() => this.handleDeleteClick(snapshot)} />}
-            <span className="icon snapshot-details-icon u-marginLeft--20 u-cursor--pointer" />
+            <Link to={`/snapshots/details/${snapshot?.name}`} className="icon snapshot-details-icon u-marginLeft--20 u-cursor--pointer" />
           </div>
         }
       </div>
@@ -61,4 +63,4 @@ class SnapshotDetailsRow extends React.Component {
   }
 }
 
-export default SnapshotDetailsRow;
+export default SnapshotRow;
