+++ kotsadm/web/src/components/snapshots/Snapshots.jsx
@@ -1,34 +1,57 @@
 import React, { Component } from "react";
-import { withRouter } from "react-router-dom"
+import { Link, withRouter } from "react-router-dom";
 import Helmet from "react-helmet";
-
 import Loader from "../shared/Loader";
-import SnapshotStorageDestination from "./SnapshotStorageDestination";
-import ConfigureSnapshots from "./ConfigureSnapshots";
+import SnapshotRow from "./SnapshotRow";
+import BackupRestoreModal from "../modals/BackupRestoreModal";
+import DeleteSnapshotModal from "../modals/DeleteSnapshotModal";
+import DummySnapshotRow from "./DummySnapshotRow";
+import GettingStartedSnapshots from "./GettingStartedSnapshots";
 
-import "../../scss/components/shared/SnapshotForm.scss";
+import "../../scss/components/snapshots/AppSnapshots.scss";
 import { Utilities } from "../../utilities/utilities";
 
 
 class Snapshots extends Component {
   state = {
+    deleteSnapshotModal: false,
+    snapshotToDelete: "",
+    deleteErr: false,
+    deleteErrorMsg: "",
+    restoreSnapshotModal: false,
+    snapshotToRestore: "",
+
     snapshotSettings: null,
     isLoadingSnapshotSettings: true,
     snapshotSettingsErr: false,
-    snapshotSettingsErrMsg: "",
-    toggleSnapshotView: false,
-    hideCheckVeleroButton: false,
-    updateConfirm: false,
-    updatingSettings: false,
-    updateErrorMsg: ""
+    snapshotSettingsErrMsg: ""
   };
 
-  fetchSnapshotSettings = (isCheckForVelero) => {
+  toggleConfirmDeleteModal = snapshot => {
+    if (this.state.deleteSnapshotModal) {
+      this.setState({ deleteSnapshotModal: false, snapshotToDelete: "", deleteErr: false, deleteErrorMsg: "" });
+    } else {
+      this.setState({ deleteSnapshotModal: true, snapshotToDelete: snapshot, deleteErr: false, deleteErrorMsg: "" });
+    }
+  };
+
+  handleDeleteSnapshot = snapshot => {
+    console.log("delete", snapshot);
+  }
+
+  toggleRestoreModal = snapshot => {
+    if (this.state.restoreSnapshotModal) {
+      this.setState({ restoreSnapshotModal: false, snapshotToRestore: "" });
+    } else {
+      this.setState({ restoreSnapshotModal: true, snapshotToRestore: snapshot });
+    }
+  };
+
+  fetchSnapshotSettings = () => {
     this.setState({
       isLoadingSnapshotSettings: true,
       snapshotSettingsErr: false,
-      snapshotSettingsErrMsg: "",
-      hideCheckVeleroButton: isCheckForVelero ? true : false
+      snapshotSettingsErrMsg: ""
     });
 
     fetch(`${window.env.API_ENDPOINT}/snapshots/settings`, {
@@ -46,13 +69,6 @@ class Snapshots extends Component {
           snapshotSettingsErr: false,
           snapshotSettingsErrMsg: "",
         })
-        if (result.veleroVersion === "") {
-          setTimeout(() => {
-            this.setState({ hideCheckVeleroButton: false });
-          }, 5000);
-        } else {
-          this.setState({ hideCheckVeleroButton: false });
-        }
       })
       .catch(err => {
         this.setState({
@@ -63,69 +79,46 @@ class Snapshots extends Component {
       })
   }
 
-  componentDidMount = () => {
+  componentDidMount() {
     this.fetchSnapshotSettings();
-  }
-
-  toggleSnapshotView = (isEmptyView) => {
-    this.setState({ toggleSnapshotView: !this.state.toggleSnapshotView, isEmptyView: isEmptyView ? isEmptyView : false });
-  }
-
-  updateSettings = (payload) => {
-    this.setState({ updatingSettings: true, updateErrorMsg: "" });
-
-    fetch(`${window.env.API_ENDPOINT}/snapshots/settings`, {
-      method: "PUT",
-      headers: {
-        "Authorization": Utilities.getToken(),
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(payload)
-    })
-      .then(async (res) => {
-
-        const settingsResponse = await res.json();
-        if (!res.ok) {
-          this.setState({
-            updatingSettings: false,
-            updateErrorMsg: settingsResponse.error
-          })
-          return;
-        }
-
-        if (settingsResponse.success) {
-          this.setState({
-            snapshotSettings: settingsResponse,
-            updatingSettings: false,
-            updateConfirm: true,
-            updateErrorMsg: ""
-          });
-          setTimeout(() => {
-            this.setState({ updateConfirm: false })
-          }, 3000);
-        } else {
-          this.setState({
-            updatingSettings: false,
-            updateErrorMsg: settingsResponse.error
-          })
-        }
-      })
-      .catch((err) => {
-        console.error(err);
-        this.setState({
-          updatingSettings: false
-        });
-      });
-  }
-
-  renderNotVeleroMessage = () => {
-    return <p className="u-color--chestnut u-fontSize--small u-fontWeight--medium u-lineHeight--normal u-marginTop--12">Not able to find Velero</p>
   }
 
 
   render() {
-    const { isLoadingSnapshotSettings, toggleSnapshotView, snapshotSettings, hideCheckVeleroButton, updateConfirm, updatingSettings, updateErrorMsg, isEmptyView } = this.state;
-    const isLicenseUpload = !!this.props.history.location.search;
+    const snapshots = [
+      {
+        appID: "1jR0VjB2Vm1lxrqoE3H0BTer6rd",
+        expiresAt: "2020-11-25T22:56:15Z",
+        finishedAt: "2020-10-26T22:56:22Z",
+        name: "qakots-g4bjh",
+        sequence: 0,
+        startedAt: "2020-10-26T22:56:15Z",
+        status: "PartiallyFailed",
+        supportBundleId: "backup-qakots-g4bjh",
+        trigger: "manual",
+        volumeBytes: 0,
+        volumeCount: 0,
+        volumeSizeHuman: "0B",
+        volumeSuccessCount: 0
+      },
+      {
+        appID: "1jR0VjB2Vm1lxrqoE3H0BTer6rk",
+        expiresAt: "2020-11-27T20:56:15Z",
+        finishedAt: "2020-10-26T20:56:22Z",
+        name: "qakots-g4bjk",
+        sequence: 0,
+        startedAt: "2020-10-25T20:56:15Z",
+        status: "Completed",
+        supportBundleId: "backup-qakots-g4bjh",
+        trigger: "manual",
+        volumeBytes: 4,
+        volumeCount: 0,
+        volumeSizeHuman: "4B",
+        volumeSuccessCount: 0
+      }
+    ]
+
+    const { isLoadingSnapshotSettings, snapshotSettings } = this.state;
 
     if (isLoadingSnapshotSettings) {
       return (
@@ -135,29 +128,55 @@ class Snapshots extends Component {
       )
     }
 
+
     return (
-      <div className="container flex-column flex1 u-overflow--auto u-paddingTop--30 u-paddingBottom--20 u-marginTop--10 alignItems--center">
+      <div className="container flex-column flex1 u-overflow--auto u-paddingTop--30 u-paddingBottom--20 alignItems--center">
         <Helmet>
           <title>Snapshots</title>
         </Helmet>
-        {(toggleSnapshotView || (snapshotSettings?.veleroVersion !== "")) ?
-          <SnapshotStorageDestination
-            snapshotSettings={snapshotSettings}
-            updateSettings={this.updateSettings}
-            fetchSnapshotSettings={this.fetchSnapshotSettings}
-            updateConfirm={updateConfirm}
-            updatingSettings={updatingSettings}
-            updateErrorMsg={updateErrorMsg}
-            renderNotVeleroMessage={this.renderNotVeleroMessage}
-            toggleSnapshotView={this.toggleSnapshotView}
-            isEmptyView={isEmptyView}
-            hideCheckVeleroButton={hideCheckVeleroButton}
-            isLicenseUpload={isLicenseUpload} /> :
-          <ConfigureSnapshots
-            snapshotSettings={snapshotSettings}
-            fetchSnapshotSettings={this.fetchSnapshotSettings}
-            renderNotVeleroMessage={this.renderNotVeleroMessage}
-            hideCheckVeleroButton={hideCheckVeleroButton} />}
+        <div className="AppSnapshots--wrapper flex1 flex-column u-width--full">
+          <div className={`flex flex-auto alignItems--flexStart justifyContent--spaceBetween ${(snapshots?.length > 0 && snapshotSettings?.veleroVersion !== "") && "u-borderBottom--gray darker"}`}>
+            <p className="u-fontWeight--bold u-color--tuna u-fontSize--larger u-lineHeight--normal u-marginBottom--15">Snapshots</p>
+            <div className="flex u-marginBottom--15">
+              <Link to={`/snapshots/settings`} className="replicated-link u-fontSize--small u-fontWeight--bold u-marginRight--20 flex alignItems--center"><span className="icon snapshotSettingsIcon u-marginRight--5" />Settings</Link>
+              <span data-for="startSnapshotBtn" data-tip="startSnapshotBtn" data-tip-disable={false}>
+                <button className="btn primary blue"> Start a snapshot</button>
+              </span>
+            </div>
+          </div>
+          {snapshots?.length > 0  && snapshotSettings?.veleroVersion !== "" ?
+            <div className="flex flex-column">
+              {snapshots?.map((snapshot) => (
+                <SnapshotRow
+                  key={`snapshot-${snapshot.name}-${snapshot.started}`}
+                  snapshot={snapshot}
+                  toggleConfirmDeleteModal={this.toggleConfirmDeleteModal}
+                  toggleRestoreModal={this.toggleRestoreModal}
+                />
+              ))}
+            </div> :
+            <div className="flex flex-column u-position--relative">
+              {[0, 1, 2, 3, 4, 5].map((el) => (<DummySnapshotRow key={el}/>
+              ))}
+              <GettingStartedSnapshots isVeleroInstalled={snapshotSettings?.veleroVersion !== ""} history={this.props.history} />
+            </div>}
+        </div>
+        {this.state.deleteSnapshotModal &&
+          <DeleteSnapshotModal
+            deleteSnapshotModal={this.state.deleteSnapshotModal}
+            toggleConfirmDeleteModal={this.toggleConfirmDeleteModal}
+            handleDeleteSnapshot={this.handleDeleteSnapshot}
+            snapshotToDelete={this.state.snapshotToDelete}
+            deleteErr={this.state.deleteErr}
+            deleteErrorMsg={this.state.deleteErrorMsg}
+          />
+        }
+        {this.state.restoreSnapshotModal &&
+          <BackupRestoreModal
+            restoreSnapshotModal={this.state.restoreSnapshotModal}
+            toggleRestoreModal={this.toggleRestoreModal}
+            snapshotToRestore={this.state.snapshotToRestore}
+          />}
       </div>
     );
   }
