+++ kotsadm/web/src/scss/components/shared/SnapshotForm.scss
@@ -1,6 +1,9 @@
 .snapshot-form-wrapper {
   width: 100%;
   max-width: 480px;
+  background: #F8F8F8;
+  border-radius: 4px;
+  padding: 10px;
 
   .BoxedCheckbox {
     padding-left: 10px;
@@ -28,4 +31,12 @@
   border-radius: 4px;
   border: 1px solid #DFDFDF;
   overflow: hidden;
+}
+
+.Info--wrapper {
+  width: 100%;
+  max-width: 480px;
+  background: #F5F8FC;
+  border-radius: 4px;
+  padding: 10px;
 }
