+++ kotsadm/web/src/scss/utilities/icons.scss
@@ -743,6 +743,16 @@
   width: 21px;
   height: 21px;
 }
+.icon.snapshot-getstarted-icon {
+  background-position: -185px -133px;
+  width: 50px;
+  height: 45px;
+}
+.icon.info-icon {
+  background-position: -225px -256px;
+  width: 20px;
+  height: 16px;
+}
 .icon.down-arrow-icon {
   background-position: -207px -230px;
   width: 12px;
