+++ kotsadm/web/src/scss/utilities/modals.scss
@@ -38,6 +38,10 @@
       max-width: 100%;
       height: 100%;
     }
+    &.ConfigureSnapshots {
+      width: 100%;
+      max-width: 570px;
+    }
     
     .Modal-header {
       padding: 20px 25px 0;
